feat: use alloy types for BlobTransactionSidecar
