+++ crates/consensus/src/transaction/envelope.rs
@@ -305,7 +305,7 @@ mod tests {
     use super::*;
     use crate::transaction::SignableTransaction;
     use alloy_eips::eip2930::{AccessList, AccessListItem};
-    use alloy_primitives::{hex, Address, Bytes, Signature, TxKind, B256, U256};
+    use alloy_primitives::{hex, Address, Bytes, Signature, TxKind, U256};
     use std::{fs, path::PathBuf, vec};
 
     #[cfg(not(feature = "std"))]
